requirements.txt
fastapi
uvicorn
pandas
openpyxl
langchain
langchain-openai
openai
pydantic
python-dotenv
pyyaml
streamlit
matplotlib
plotly
prometheus-client
pytest
httpx
